requirements.txt
absl-py==2.1.0
aiohttp==3.9.4
aioresponses==0.7.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.1
astunparse==1.6.3
attrs==23.2.0
awscli==1.29.54
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.28.23
botocore==1.31.54
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.0
dataclasses-json==0.6.4
deepface==0.0.89
defusedxml==0.7.1
distro==1.9.0
docker==6.1.3
docutils==0.16
fastapi==0.110.1
filelock==3.13.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
gdown==5.1.0
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.62.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
imageio==2.34.0
inference==0.9.17
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.2.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
mtcnn==0.1.1
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.7
networkx==3.3
numpy==1.25.2
onnxruntime==1.15.1
openai==1.17.1
opencv-python==4.8.0.76
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pendulum==3.0.0
piexif==1.1.3
pillow==10.3.0
pluggy==1.4.0
prettytable==3.10.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.20.0
protobuf==4.25.3
psutil==5.9.8
PuLP==2.8.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pybase64==1.3.2
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.1.1
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
requests==2.31.0
requests-toolbelt==1.0.0
retina-face==0.0.16
rich==13.5.2
rpds-py==0.18.0
rsa==4.7.2
s3transfer==0.6.2
scikit-image==0.23.1
scipy==1.13.0
seaborn==0.13.2
shapely==2.0.1
six==1.16.0
skypilot==0.4.1
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
structlog==24.1.0
supervision==0.19.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tf_keras==2.16.0
thop==0.1.1.post2209072238
tifffile==2024.2.12
time-machine==2.14.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ultralytics==8.1.47
urllib3==1.26.18
wcwidth==0.2.13
websocket-client==1.7.0
Werkzeug==3.0.2
wrapt==1.16.0
yarl==1.9.4
